requirements.txt
streamlit
openai
python-dotenv
faiss-cpu
sentence-transformers
langchain
langchain-community

